messaging_app/requirements-test.txt
# Test dependencies for CI/CD pipeline
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
coverage==7.6.1
factory-boy==3.3.1
mixer==7.2.2

# Code quality tools
flake8==7.1.1
black==24.8.0
isort==5.13.2

# Additional testing utilities
requests==2.32.3
responses==0.25.3
freezegun==1.5.1

# Documentation
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
